+++ docs/src/routes/demo/samples/highlightjs.svelte
@@ -4,13 +4,10 @@
 
     hljs.registerLanguage("html", xml);
 
-    const highlightElement = (element, syntax) => hljs.highlightElement(element);
-
-    const highlightCode = (code, syntax) =>
+    const highlight = (code, syntax) =>
         hljs.highlight(code, {
             language: syntax,
         }).value;
-
 </script>
 
 <script>
@@ -25,11 +22,13 @@
 
     const CodeJar = $store;
 
-    let syntax = DEFAULT_CODE_SYNTAX;
-    let value = DEFAULT_CODE_SAMPLE;
-
     $: ({withLineNumbers = false} = $options);
-
 </script>
 
-<CodeJar {highlightCode} {highlightElement} {syntax} {withLineNumbers} bind:value />
+<CodeJar
+    class="hljs"
+    syntax={DEFAULT_CODE_SYNTAX}
+    value={DEFAULT_CODE_SAMPLE}
+    {highlight}
+    {withLineNumbers}
+/>